Modularize periods
